Push AliasChooser implementation down to leaf classes.
Moving AliasChooser up to AbstractConscryptSocket was a mistake
as it also adds it to OpenSSLSocketImpl which is unwanted.

Pushing it back down to the leaf socket implementations causes
some minor code duplication but this will go away with
ConscryptFileDescriptorSocket.
