+++ common/src/main/java/org/conscrypt/AbstractConscryptSocket.java
@@ -39,12 +39,11 @@ import javax.net.ssl.SSLSession;
 import javax.net.ssl.SSLSocket;
 import javax.net.ssl.X509KeyManager;
 import javax.security.auth.x500.X500Principal;
-import org.conscrypt.SSLParametersImpl.AliasChooser;
 
 /**
  * Abstract base class for all Conscrypt {@link SSLSocket} classes.
  */
-abstract class AbstractConscryptSocket extends SSLSocket implements AliasChooser {
+abstract class AbstractConscryptSocket extends SSLSocket {
     final Socket socket;
     private final boolean autoClose;
 
@@ -513,17 +512,6 @@ abstract class AbstractConscryptSocket extends SSLSocket implements AliasChooser
             builder.append(super.toString());
         }
         return builder.toString();
-    }
-
-    @Override
-    public final String chooseServerAlias(X509KeyManager keyManager, String keyType) {
-        return keyManager.chooseServerAlias(keyType, null, this);
-    }
-
-    @Override
-    public final String chooseClientAlias(X509KeyManager keyManager, X500Principal[] issuers,
-        String[] keyTypes) {
-        return keyManager.chooseClientAlias(keyTypes, issuers, this);
     }
 
     /**
